Store StumbleBundles in the ContentProvider.
This moves the code that creates a ContentValues object out of the
Reporter and into the MainActivity. When a StumblerBundle is ready,
the library will broadcast it to the MainActivity, which will turn
it into a JSON object and store it into the ContentProvider to be
sent later.

Known bugs/FIXMEs:
- Relying on MainActivity to receive a broadcast is wrong, as the
onStop() method unregisters the activity from all broadcasts. There
should be a separate BroadcastReceiver that registers via the manifest
for this.

- The SyncAdapter still knows about the format of the JSON the server
expects, since it needs to wrap a bunch of (JSON representations of)
StumblerBundles into one wrapper object and an "items" array before
uploading. We should at the very least provide a helper method for
that in the library: StumblerUtils.formatForMLS(StumblerBundle[]).

- The location data goes through some seemingly redundant conversions
before being uploaded. Basically a StumblerBundle reaches MainActivity,
which converts it to JSON, then inserts the individual fields of the
JSON object into the ContentProvider (wifi and cell information are
simply stored as JSON strings), from where the SyncAdapter reads them
and converts back do JSON. It's probably possible to just store a JSON
string in the ContentProvider instead.