The grand refactor to remove the DB. - Delete ContentResolver(s), Database.java
- DatabaseContract is now DataStorageContract, and has the constants
  used for file storage

- DataStorageManager is the backend, it has an in-memory working buffer
  for collecting reports, and in-memory buffer of reports to be sent,
and a list of files on disk (with about 50 reports per file). Please see the documentation  for the class, it explains this in more detail.

-- After data leaves the in-memory working buffer, it is zipped, this
required some changes to the network code to handle pre-zipped data
(AbstractCommunicator.java)

-- The constructor of DataStorageManager also upgrades the sync stats from the old DB.

- Removing the DB propagates changes to StumblerBundle, StumblerBundleReceiver, Submitter, Reporter.
The logic is unchanged in these classes, just cleanup for the backend.