See Fennec integration bug https://bugzilla.mozilla.org/show_bug.cgi?id=1038843
- Refactored StumblerService, moved bound usage to ClientStumblerService,
this is not used by Fennec.
- Restructure code by thread. Concurrency bugs are the highest-risk, and
the code can be clearly organized by thread so future contributers can
see at a glance where thread divisions are.
- ensure that stumbler service init() happens on the correct thread, in
MozStumbler, this is the main thread, in the Fennec service, this is on
the service's dedicated thread
- remove the use of the blocklists from Fennec.
- changed uploading in fennec mode to only upload after a start intent.
Need to ensure that when foregrounded, Fennec sends this message.