This reworks the LRUMapTileCache to use a proper LRU.
The capacity of the underlying LinkedHashMap has been modified to be 2x the requested capacity as there's a mismatch in meanings.

The callers is asking for capacity in terms of tiles, the capacity for a linked hash map is the # of hash buckets available which is a smaller number.

The final effect of this patch is that memory usage is pretty constant @ ~37MB on a Moto G with a 5x5 tile screen, but this adds significant flicker.

The upside is that this finally seems to fix the memory consumption error and no more OOM death.