* Dropped ZoomInTileLooper and ZoomOutTileLooper entirely as they no longer do anything. * Added an invalidate() call on MapView::setZoom so that any zoom   change forces a viewport redraw
Note that this will snap back your zoom level if you haven't zoomed in
by a full level.  This behaviour was less obvious before the patch,
but is happened before as well.  It just happened in 2 steps.  Pinch
zoom with an edge that is not redraw, then on the next redraw, the
zoom level would snap back.