Motion Detect: had an extra wait period for first GPS fix. This logic is of no value. . Remove this extra time window, and just use the existing time window, called mPrefMotionChangeTimeWindowMs, in this code. The code is cleaner with one time window (as it now has) instead of 2.
The reason the old code was not good, is that the extra time window to
wait for GPS fix is not any different than the existing motion change
time window, except that it was hard coded to 5 mins.

What would happen is that when scanning was started, and there is no GPS fix,
it would wait 5 mins before the motion detection logic would kick in.
This is confusing to anyone trying to understand the workings
of this code. It should be easy to grok based on the developer setting
(the setting for wait XXX seconds before pausing).