This fixes a long standing oddity that IHttpUtil was passed in as a reference sometimes, and instantiated raw other times.
Now we just grab it from the ServiceLocator.

The UpdaterTest has been simplified to overload the mock networking
using the ServiceLocator.

Minor changes to ServiceConfig are included so that only strings are
needed to instantiate classes, so that the ServiceConfig class does
not need build time dependencies on any other class.