Update Travis CI build status image in README.md (#1801)
* Link Travis build status image to Travis status page

Instead of linking to the logo image, link to the Travis CI page
for the project, which provides information about the build status.
.

* Show the Travis build status for the default branch

The logo currently displays the build status of the last build from any
branch. Change it to display the last build from from the default branch
(dev).